refactor: move feature classes to datasets module
